src/pyproject.toml
[project]
name = "litestar-crud"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "litestar[standard] (>=2.16.0,<3.0.0)",
    "granian (>=2.2.6,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "advanced-alchemy (>=1.4.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
]

[tool.poetry]
packages = [{include = "litestar_crud", from = "src"}]


[tool.poetry.group.dev.dependencies]
pudb = "^2025.1"
mypy = "^1.15.0"
black = "^25.1.0"
ptpython = "^3.0.30"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
